+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/robots/legged_robot/legged_robot_cfg.py
@@ -31,7 +31,12 @@ class LeggedRobotCfg(RobotBaseCfg):
         """Physics material applied to the feet of the robot."""
 
         prim_path = "/World/Materials/footMaterial"
-        """Path to the physics material prim. Default: /World/Materials/footMaterial."""
+        """Path to the physics material prim. Default: /World/Materials/footMaterial.
+
+        Note:
+            If the prim path is not absolute, it will be resolved relative to the path specified when spawning
+            the object.
+        """
         static_friction: float = 1.0
         """Static friction coefficient. Default: 1.0."""
         dynamic_friction: float = 1.0